Implement a cancelable read via daemon thread abandonment
Since there is no way to make the handle received (as fd
0/STD_INPUT_HANDLE) refer to an object with FILE_FLAG_OVERLAPPED, we
really have no way to actually _interrupt_ a read request once it is
issued.  Our second-best option, implemented here is to do the read
operation in a background thread that can be abandoned to exit if such a
condition is detected.